+++ src/shared/mod.rs
@@ -1,5 +1,7 @@
 pub use byte_range::*;
+pub use hexstring::*;
 pub use recording::*;
 
 mod byte_range;
+mod hexstring;
 mod recording;