+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fixed an issue where `#EXT-X-MAP` and `#EXT-X-KEY` files were re-downloaded when they re-appear on a later segment.
+
 ## v0.1.0 (2025-10-22)
 
 - Initial release.